Support first init without packaged testdrivers

+++ src/main/java/de/interactive_instruments/etf/webapp/controller/EtfConfigController.java
@@ -487,19 +487,21 @@ public class EtfConfigController implements PropertyHolder {
 		// Copy test drivers
 		final String tdDirName = "/testdrivers";
 		final Set<String> tds = servletContext.getResourcePaths(tdDirName);
-		for (final String td : tds) {
-			final String testDriverName = td.substring(tdDirName.length());
-			if (!SUtils.isNullOrEmpty(testDriverName) && latestDriverVersions.isNewer(testDriverName)) {
-				logger.info("Installing Test Driver " + testDriverName);
-				final IFile tdJar = new IFile(tdDir, testDriverName);
-				final InputStream jarStream = servletContext.getResourceAsStream(td);
-				try (final FileOutputStream out = new FileOutputStream(tdJar)) {
-					IOUtils.copy(jarStream, out);
-				} catch (final IOException e) {
-					tdJar.delete();
-					logger.error("Could not copy test driver: ", e);
-				} finally {
-					jarStream.close();
+		if (tds != null) {
+			for (final String td : tds) {
+				final String testDriverName = td.substring(tdDirName.length());
+				if (!SUtils.isNullOrEmpty(testDriverName) && latestDriverVersions.isNewer(testDriverName)) {
+					logger.info("Installing Test Driver " + testDriverName);
+					final IFile tdJar = new IFile(tdDir, testDriverName);
+					final InputStream jarStream = servletContext.getResourceAsStream(td);
+					try (final FileOutputStream out = new FileOutputStream(tdJar)) {
+						IOUtils.copy(jarStream, out);
+					} catch (final IOException e) {
+						tdJar.delete();
+						logger.error("Could not copy test driver: ", e);
+					} finally {
+						jarStream.close();
+					}
 				}
 			}
 		}
